+++ src/ascii2d.js
@@ -1,11 +1,11 @@
 import _ from 'lodash';
 import * as Cheerio from 'cheerio';
 import FormData from 'form-data';
-import pixivShorten from './urlShorten/pixiv';
 import logError from './logError';
-import { retryAync } from './utils/retry';
-import { getCqImg64FromUrl } from './utils/image';
 import CQ from './CQcode';
+import { retryAsync } from './utils/retry';
+import { getCqImg64FromUrl, getAntiShieldedCqImg64FromUrl } from './utils/image';
+import { confuseURL } from './utils/url';
 const Axios = require('./axiosProxy');
 
 let hostsI = 0;
@@ -22,12 +22,14 @@ async function doSearch(url, snLowAcc = false) {
   if (host === 'ascii2d.net') host = `https://${host}`;
   else if (!/^https?:\/\//.test(host)) host = `http://${host}`;
   const callApi = global.config.bot.ascii2dLocalUpload ? callAscii2dUploadApi : callAscii2dUrlApi;
-  const { colorURL, colorDetail } = await retryAync(
+  const { colorURL, colorDetail } = await retryAsync(
     async () => {
       const ret = await callApi(host, url);
       const colorURL = ret.request.res.responseUrl;
       if (!colorURL.includes('/color/')) {
         const $ = Cheerio.load(ret.data, { decodeEntities: false });
+        console.error('[error] ascii2d url:', colorURL);
+        console.error($('body').text());
         throw new Error($('.container > .row > div:first-child > p').text().trim());
       }
       return {
@@ -71,23 +73,34 @@ function getDetail(ret, baseURL) {
   let result = {};
   const html = ret.data;
   const $ = Cheerio.load(html, { decodeEntities: false });
-  const $itembox = $('.item-box');
-  for (let i = 0; i < $itembox.length; i++) {
-    const $box = $($itembox[i]);
+  const $itemBox = $('.item-box');
+  for (let i = 0; i < $itemBox.length; i++) {
+    const $box = $($itemBox[i]);
     const $link = $box.find('.detail-box a');
-    if ($link.length === 0) continue;
-    const $title = $($link[0]);
-    const $author = $($link[1]);
-    result = {
-      thumbnail: baseURL + $box.find('.image-box img').attr('src'),
-      title: $title.html(),
-      author: $author.html(),
-      url: $title.attr('href'),
-      author_url: $author.attr('href'),
-    };
-    break;
+    // 普通结果
+    if ($link.length) {
+      const $title = $($link[0]);
+      const $author = $($link[1]);
+      result = {
+        thumbnail: baseURL + $box.find('.image-box img').attr('src'),
+        title: $title.text(),
+        author: $author.text(),
+        url: $title.attr('href'),
+        author_url: $author.attr('href'),
+      };
+      break;
+    }
+    // 人为提交结果
+    const $external = $box.find('.external');
+    if ($external.length) {
+      result = {
+        thumbnail: baseURL + $box.find('.image-box img').attr('src'),
+        title: $external.text(),
+      };
+      break;
+    }
   }
-  if (!result.url) {
+  if (!result.title) {
     logError(`${global.getTime()} [error] ascii2d getDetail`);
     logError(ret);
   }
@@ -95,13 +108,15 @@ function getDetail(ret, baseURL) {
 }
 
 async function getResult({ url, title, author, thumbnail, author_url }, snLowAcc = false) {
-  if (!url) return { success: false, result: '由未知错误导致搜索失败' };
-  const texts = [CQ.escape(`「${title}」/「${author}」`)];
+  if (!title) return { success: false, result: '由未知错误导致搜索失败' };
+  const texts = [CQ.escape(author ? `「${title}」/「${author}」` : title)];
   if (thumbnail && !(global.config.bot.hideImg || (snLowAcc && global.config.bot.hideImgWhenLowAcc))) {
-    texts.push(await getCqImg64FromUrl(thumbnail));
+    const mode = global.config.bot.antiShielding;
+    if (mode > 0) texts.push(await getAntiShieldedCqImg64FromUrl(thumbnail, mode));
+    else texts.push(await getCqImg64FromUrl(thumbnail));
   }
-  texts.push(CQ.escape(pixivShorten(url)));
-  if (author_url) texts.push(`Author: ${CQ.escape(pixivShorten(author_url))}`);
+  if (url) texts.push(CQ.escape(confuseURL(url)));
+  if (author_url) texts.push(`Author: ${CQ.escape(confuseURL(author_url))}`);
   return { success: true, result: texts.join('\n') };
 }