+++ src/logError.js
@@ -1,11 +1,21 @@
+import _ from 'lodash';
+import { AxiosError } from 'axios';
+
 export default e => {
-  if (typeof e === 'object' && e.stack) {
-    console.error(e.stack);
-    delete e.stack;
-    if (e instanceof Error) {
-      console.error(JSON.stringify(e, Object.getOwnPropertyNames(e), 2));
-      return;
-    }
+  if (e instanceof AxiosError) {
+    console.error(String(e.stack || e));
+    console.error({
+      config: e.config ? _.pick(e.config, ['method', 'url', 'data']) : undefined,
+      response: e.response
+        ? {
+            ..._.pick(e.response, ['status', 'statusText']),
+            ...((e.response.headers['content-type'] || '').includes('application/json')
+              ? _.pick(e.response, ['data'])
+              : {}),
+          }
+        : undefined,
+    });
+    return;
   }
   console.error(e);
 };