+++ src/plugin/bilibili/push.js
@@ -4,9 +4,11 @@ import emitter from '../../emitter';
 import logError from '../../logError';
 import { sleep } from '../../utils/sleep';
 import { getUserNewDynamicsInfo } from './dynamic';
-import { getUserLiveData } from './live';
+import { getUsersLiveData } from './live';
+import { getUserSeasonNewVideosInfo } from './season';
+import { purgeLink } from './utils';
 
-let pushConfig = { dynamic: {}, live: {} };
+let pushConfig = { dynamic: {}, live: {}, season: {}, series: {} };
 const liveStatusMap = new Map();
 let checkPushTask = null;
 
@@ -21,7 +23,7 @@ function init() {
   for (const uid of liveStatusMap.keys()) {
     if (!(uid in pushConfig.live)) liveStatusMap.delete(uid);
   }
-  if (_.size(pushConfig.dynamic) || _.size(pushConfig.live)) {
+  if (_.size(pushConfig.dynamic) || _.size(pushConfig.live) || _.size(pushConfig.season)) {
     checkPushTask = setInterval(checkPush, Math.max(global.config.bot.bilibili.pushCheckInterval, 30) * 1000);
     checkPush();
   }
@@ -30,6 +32,8 @@ function init() {
 function getPushConfig() {
   const dynamic = {};
   const live = {};
+  const season = {};
+  const series = {};
   _.each(global.config.bot.bilibili.push, (confs, uid) => {
     if (!Array.isArray(confs)) return;
     dynamic[uid] = [];
@@ -40,13 +44,28 @@ function getPushConfig() {
         live[uid].push({ gid: conf });
       } else if (typeof conf === 'object' && typeof conf.gid === 'number') {
         if (conf.dynamic === true) dynamic[uid].push({ gid: conf.gid, atAll: conf.dynamicAtAll });
+        else if (conf.video === true) dynamic[uid].push({ gid: conf.gid, atAll: conf.dynamicAtAll, onlyVideo: true });
         if (conf.live === true) live[uid].push({ gid: conf.gid, atAll: conf.liveAtAll });
+        if (conf.seasons && conf.seasons.length) {
+          conf.seasons.forEach(sid => {
+            const key = `${uid}:${sid}`;
+            if (!season[key]) season[key] = [];
+            season[key].push({ gid: conf.gid, atAll: conf.seasonAtAll });
+          });
+        }
+        if (conf.series && conf.series.length) {
+          conf.series.forEach(sid => {
+            const key = `${uid}:${sid}`;
+            if (!series[key]) series[key] = [];
+            series[key].push({ gid: conf.gid, atAll: conf.seriesAtAll });
+          });
+        }
       }
     });
     if (!dynamic[uid].length) delete dynamic[uid];
     if (!live[uid].length) delete live[uid];
   });
-  return { dynamic, live };
+  return { dynamic, live, season, series };
 }
 
 async function checkPush() {
@@ -59,6 +78,16 @@ async function checkPush() {
       }),
       checkLive().catch(e => {
         logError(`${global.getTime()} [error] bilibili check live`);
+        logError(e);
+        return [];
+      }),
+      checkSeason('season').catch(e => {
+        logError(`${global.getTime()} [error] bilibili check season`);
+        logError(e);
+        return [];
+      }),
+      checkSeason('series').catch(e => {
+        logError(`${global.getTime()} [error] bilibili check series`);
         logError(e);
         return [];
       }),
@@ -81,10 +110,11 @@ async function checkDynamic() {
   for (const [uid, confs] of Object.entries(pushConfig.dynamic)) {
     const dynamics = dynamicMap[uid];
     if (!dynamics || !dynamics.length) continue;
-    for (const dynamic of dynamics) {
-      for (const { gid, atAll } of confs) {
+    for (const { type, text } of dynamics) {
+      for (const { gid, atAll, onlyVideo } of confs) {
+        if (onlyVideo && type !== 8) continue;
         tasks.push(() =>
-          global.sendGroupMsg(gid, atAll ? `${dynamic}\n\n${CQ.atAll()}` : dynamic).catch(e => {
+          global.sendGroupMsg(gid, atAll ? `${text}\n\n${CQ.atAll()}` : text).catch(e => {
             logError(`${global.getTime()} [error] bilibili push dynamic to group ${gid}`);
             logError(e);
           })
@@ -96,12 +126,7 @@ async function checkDynamic() {
 }
 
 async function checkLive() {
-  const liveMap = {};
-  await Promise.all(
-    Object.keys(pushConfig.live).map(async uid => {
-      liveMap[uid] = await getUserLiveData(uid);
-    })
-  );
+  const liveMap = await getUsersLiveData(Object.keys(pushConfig.live));
   const tasks = [];
   for (const [uid, confs] of Object.entries(pushConfig.live)) {
     const liveData = liveMap[uid];
@@ -109,15 +134,46 @@ async function checkLive() {
     const { status, name, url, title, cover } = liveData;
     const oldStatus = liveStatusMap.get(uid);
     liveStatusMap.set(uid, status);
-    if (status && !oldStatus) {
+    if (status === 1 && status !== oldStatus) {
       for (const { gid, atAll } of confs) {
         tasks.push(() =>
           global
-            .sendGroupMsg(gid, [CQ.img(cover), `【${name}】${title}`, url, ...(atAll ? [CQ.atAll()] : [])].join('\n'))
+            .sendGroupMsg(
+              gid,
+              [CQ.img(cover), `【${name}】${title}`, purgeLink(url), ...(atAll ? [CQ.atAll()] : [])].join('\n')
+            )
             .catch(e => {
               logError(`${global.getTime()} [error] bilibili push live status to group ${gid}`);
               logError(e);
             })
+        );
+      }
+    }
+  }
+  return tasks;
+}
+
+/**
+ * @param {'season' | 'series'} type
+ */
+async function checkSeason(type) {
+  const map = {};
+  await Promise.all(
+    Object.keys(pushConfig[type]).map(async usid => {
+      map[usid] = await getUserSeasonNewVideosInfo(usid, type);
+    })
+  );
+  const tasks = [];
+  for (const [usid, confs] of Object.entries(pushConfig[type])) {
+    const texts = map[usid];
+    if (!texts || !texts.length) continue;
+    for (const text of texts) {
+      for (const { gid, atAll } of confs) {
+        tasks.push(() =>
+          global.sendGroupMsg(gid, atAll ? `${text}\n\n${CQ.atAll()}` : text).catch(e => {
+            logError(`${global.getTime()} [error] bilibili push ${type} video to group ${gid}`);
+            logError(e);
+          })
         );
       }
     }