+++ src/saucenao.js
@@ -1,10 +1,10 @@
 import _ from 'lodash';
 import nhentai from './nhentai';
 import getSource from './getSource';
-import pixivShorten from './urlShorten/pixiv';
 import logError from './logError';
-import { getCqImg64FromUrl } from './utils/image';
 import CQ from './CQcode';
+import { getAntiShieldedCqImg64FromUrl, getCqImg64FromUrl } from './utils/image';
+import { confuseURL } from './utils/url';
 const Axios = require('./axiosProxy');
 
 let hostsI = 0;
@@ -169,7 +169,7 @@ async function doSearch(imgURL, db, debug = false) {
           // 如果是本子
           const doujinName = jp_name || eng_name; // 本子名
           if (doujinName) {
-            if (global.config.bot.getDojinDetailFromNhentai) {
+            if (global.config.bot.getDoujinDetailFromNhentai) {
               const searchName = (eng_name || jp_name).replace('(English)', '').replace(/_/g, '/');
               const doujin = await nhentai(searchName).catch(e => {
                 logError(`${global.getTime()} [error] nhentai`);
@@ -233,24 +233,16 @@ async function doSearch(imgURL, db, debug = false) {
   };
 }
 
-/**
- * 链接混淆
- *
- * @param {string} url
- * @returns
- */
-async function confuseURL(url) {
-  return pixivShorten(url);
-}
-
 async function getShareText({ url, title, thumbnail, author_url, source }) {
   const texts = [title];
   if (thumbnail && !global.config.bot.hideImg) {
-    texts.push(await getCqImg64FromUrl(thumbnail));
+    const mode = global.config.bot.antiShielding;
+    if (mode > 0) texts.push(await getAntiShieldedCqImg64FromUrl(thumbnail, mode));
+    else texts.push(await getCqImg64FromUrl(thumbnail));
   }
-  if (url) texts.push(await confuseURL(url));
-  if (author_url) texts.push(`Author: ${await confuseURL(author_url)}`);
-  if (source) texts.push(`Source: ${await confuseURL(source)}`);
+  if (url) texts.push(confuseURL(url));
+  if (author_url) texts.push(`Author: ${confuseURL(author_url)}`);
+  if (source) texts.push(confuseURL(source));
   return texts.join('\n');
 }
 
@@ -269,7 +261,7 @@ function getSearchResult(host, api_key, imgURL, db = 999) {
   return Axios.get(`${host}/search.php`, {
     params: {
       ...(api_key ? { api_key } : {}),
-      db: db,
+      db,
       output_type: 2,
       numres: 3,
       url: imgURL,