+++ CHANGELOG.md
@@ -2,6 +2,97 @@
 
 ## 2022
 
+### 11-19 v2.42.0
+
+- 将 `pixiv.net` 加入红链处理名单（否则有概率被屏蔽）
+- 使用新的红链处理方式，可以发出完整链接，只需手动复制到浏览器打开即可，不需要手动删除字符，大部分主流浏览器都可以正常打开（Chrome/Safari/Edge等）
+- 如需照顾少数无法正常打开链接的浏览器，请设置 `bot.handleBannedHostsWithLegacyMethod` 为 `true`
+- 为方便复制链接，移除 `Source:` 文字
+- 配置项变更
+  - A `bot.handleBannedHostsWithLegacyMethod`
+
+### 11-13 v2.41.0
+
+- 哔哩哔哩推送支持视频列表（以前只支持视频合集，配置见 wiki）[#373](../../issues/373)
+
+### 11-07 v2.40.2
+
+- 修复哔哩哔哩动态内链接的处理问题 [#370](../../issues/370)
+- 根据哔哩哔哩推送检测间隔过滤掉过旧的动态以免B站 API 抽风导致重复推送 [#370](../../issues/370)
+
+### 10-31 v2.40.1
+
+- 修复哔哩哔哩用户没有动态时添加推送会不停打印错误日志的问题 [#369](../../issues/369)
+- 修复 `bot.handleBannedHosts` 配置没有生效的问题
+
+### 10-29 v2.40.0
+
+- 支持输出 ascii2d 人为提交的搜索结果 [#361](../../issues/361)
+- 哔哩哔哩解析支持当原消息撤回时同步撤回解析结果（默认启用，可在配置中关闭）
+- 配置项变更
+  - A `bot.bilibili.respondRecall`
+
+### 09-27 v2.39.6
+
+- 修复有时候输出错误日志会发生额外错误的问题
+
+### 09-27 v2.39.5
+
+- 修复某些情况下 bot 会崩溃的问题
+
+### 09-22 v2.39.4
+
+- 修复哔哩哔哩直播会推送轮播的问题
+
+### 09-22 v2.39.3
+
+- 修复了哔哩哔哩直播推送失效的问题 [#360](../../issues/359)
+- 搜图结果合并转发中附带的原图也会按照 `bot.antiShielding` 进行反和谐处理
+
+### 09-10 v2.39.2
+
+- 优化了方舟公招计算识别，现在不需要特地选中高资或资深了，大概 [#359](../../issues/359)
+- 修复了方舟公招计算高资和资深提示背景异常的问题
+
+### 09-09 v2.39.1
+
+- 过滤直播推送链接参数 [#358](../../issues/358)
+- 优化错误日志输出
+
+### 08-28 v2.39.0
+
+- 图片反和谐支持组合方式，详见配置注释
+- 支持对搜图结果缩略图进行反和谐
+- 目前发现搜图结果中包含红链会导致消息无法发出，因此默认会对红链进行处理，可在配置中关闭
+- nHentai 上了 cf 五秒盾导致无法搜索本子，因此 `bot.getDoujinDetailFromNhentai` 默认值改为 `false`，并建议关闭
+- 配置项变更
+  - A `bot.antiShielding`
+  - A `bot.handleBannedHosts`
+  - M `bot.getDoujinDetailFromNhentai`
+  - M `bot.setu.antiShielding`
+
+### 07-21 v2.38.0
+
+- 由于历史原因，机器人硬编码忽略了 `/` 与 `<` 开头的发言以兼容同时部署了其他机器人的情况，现将这一功能提至配置项 `bot.regs.ignore`，默认为 `""` 即不进行忽略，如有需要请在更新后设置此项 [#344](../../issues/344)
+- 配置项变更
+  - A `bot.regs.ignore`
+
+### 07-11 v2.37.2
+
+- 哔哩哔哩解析被删除的动态或视频时增加回复提示 [#336](../../issues/336)
+- 更改部分功能使用到的 jsdelivr 域名防止国内部分地区无法访问 cdn.jsdelivr.net => fastly.jsdelivr.net
+
+### 07-10 v2.37.1
+
+- 修复哔哩哔哩动态解析问题 [#339](../../issues/339)
+
+### 07-08 v2.37.0
+
+- 哔哩哔哩推送动态支持只推送视频 [#335](../../issues/335)
+- 哔哩哔哩推送支持视频合集 [#335](../../issues/335)
+- 配置项变更
+  - M `bot.bilibili.push`（见 wiki）
+
 ### 07-04 v2.36.1
 
 - 哔哩哔哩动态解析支持投票 [#333](../../issues/333)
@@ -158,7 +249,7 @@
 
 ### 10-24 v2.30.3
 
-- 因 pixiv.cat 在大陆被墙，`bot.setu.sendPximgProxys` 默认配置变更为 `["https://i.pixiv.re/"]`；设置了 `bot.setu.pximgProxy` 的用户也请注意修改
+- 因 pixiv.cat 在大陆被墙，`bot.setu.sendPximgProxies` 默认配置变更为 `["https://i.pixiv.re/"]`；设置了 `bot.setu.pximgProxy` 的用户也请注意修改
 - setu 使用在线反代时，如果设置了代理，反和谐会走代理下载图片；不反和谐时仍由 go-cqhttp 下载图片，这种情况下如果需要走代理，需要给 go-cqhttp 配置代理
 
 ### 10-24 v2.30.2
@@ -238,7 +329,7 @@
 - whatanime 发送预览视频不再需要依赖 ffmpeg
 - 改进方舟公招计算器的识别逻辑
 - 默认 OCR 服务变更为为 qq
-- `bot.setu.pximgProxy` 和 `bot.setu.sendPximgProxys` 新增支持一些占位符
+- `bot.setu.pximgProxy` 和 `bot.setu.sendPximgProxies` 新增支持一些占位符
 - 配置项变更
   - M `bot.ocr.use` 默认值 `"ocr.space"` -> `"qq"`
   - M `bot.akhr.ocr` 默认值 `"ocr.space"` -> `"qq"`
@@ -309,7 +400,7 @@ git pull
 - 配置项变更
   - M `bot.setu.pximgServerPort` 默认值 `60233` -> `0`
   - A `bot.pmSearchResultTemp`
-  - A `bot.setu.sendPximgProxys`
+  - A `bot.setu.sendPximgProxies`
   - A `bot.setu.shortenPximgProxy`
   - A `bot.setu.r18OnlyUrl`
   - A `bot.setu.r18OnlyPrivate`
@@ -402,7 +493,7 @@ git pull
 - `ocr_image` API 移除实验模式，如有使用请将 go-cqhttp 升级到 v0.9.34 以上
 - 可设置 saucenao 在搜到本子时是否进一步去 nhentai 搜索（默认是，与之前的行为一致，若没有特殊需求不需要动该设置）([#134](../../issues/134))
 - 配置项变更
-  - A `bot.getDojinDetailFromNhentai`
+  - A `bot.getDoujinDetailFromNhentai`
 
 ### 01-25 v2.21.9